Correção no metodo de combate, ajustes nos prints do codigo

+++ JogoPOO/src/SistemaInventario/Arma.java
@@ -4,11 +4,11 @@ public class Arma extends Item{
 	
 	//atributos
 	
-	public int dano; //dano da arma
+	public double dano; //dano da arma
 	
 	//construtor
 	
-	public Arma(String nome_Item, String descricao_Item, int dano) {
+	public Arma(String nome_Item, String descricao_Item, double dano) {
 		this.nome_Item = nome_Item;
 		this.descricao_Item = descricao_Item;
 		this.dano = dano;
@@ -18,11 +18,11 @@ public class Arma extends Item{
 
 	//getters and setters
 	
-	public int getDano() {
+	public double getDano() {
 		return dano;
 	}
 
-	public void setDano(int dano) {
+	public void setDano(double dano) {
 		this.dano = dano;
 	}
